+++ contracts/ecosystem/YodaTimelock.sol
@@ -16,7 +16,7 @@ contract YodaTimelock is TimelockControllerUpgradeable, UUPSUpgradeable {
     /// @dev AccessControl Upgrader Role
     bytes32 internal constant UPGRADER_ROLE = keccak256("UPGRADER_ROLE");
     /// @dev UUPS version tracker
-    uint8 public version;
+    uint32 public version;
     uint256[50] private __gap;
 
     /**
